requirements.txt
Django==1.9.6
django-daterange-filter==1.3.0
Pillow==3.2.0
pytz==2016.4
